+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.2"
     defaultConfig {
-        applicationId "com.example.genericsl"
+        applicationId "com.genericsl"
         minSdkVersion 15
         targetSdkVersion 29
         versionCode 1
