--- app/src/main/java/com/example/genericsl/model/IUser.kt
@@ -1,8 +1,0 @@
-package com.example.genericsl.model
-
-interface IUser {
-    val email:String
-    val password:String
-    fun isDataValid():Int
-    //val isDataValid:Boolean
-}
