--- app/src/main/java/com/example/genericsl/presenter/ILoginPresenter.kt
@@ -1,7 +1,0 @@
-package com.example.genericsl.presenter
-
-interface ILoginPresenter {
-
-    fun onLogin(email:String, password:String)
-
-}
