--- app/src/main/java/com/example/genericsl/view/ui/activity/ILoginView.kt
@@ -1,10 +1,0 @@
-package com.example.genericsl.view.ui.activity
-
-interface ILoginView {
-
-    fun onLoginSuccess(message:String)
-    fun onLoginError(message:String)
-
-    //fun onLoginResult(message:String)
-
-}
