--- app/src/main/java/com/example/genericsl/view/ui/fragment/send/SendViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.genericsl.view.ui.fragment.send
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class SendViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is send Fragment"
-    }
-    val text: LiveData<String> = _text
-}
