--- app/src/main/java/com/example/genericsl/view/ui/fragment/share/ShareViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.genericsl.view.ui.fragment.share
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ShareViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is share Fragment"
-    }
-    val text: LiveData<String> = _text
-}
