+++ app/src/main/java/com/genericsl/model/IUser.kt
@@ -1,0 +1,8 @@
+package com.genericsl.model
+
+interface IUser {
+    val email:String
+    val password:String
+    fun isDataValid():Int
+    //val isDataValid:Boolean
+}
