+++ app/src/main/java/com/genericsl/presenter/ILoginPresenter.kt
@@ -1,0 +1,7 @@
+package com.genericsl.presenter
+
+interface ILoginPresenter {
+
+    fun onLogin(email:String, password:String)
+
+}
