+++ app/src/main/java/com/genericsl/view/ui/activity/ILoginView.kt
@@ -1,0 +1,10 @@
+package com.genericsl.view.ui.activity
+
+interface ILoginView {
+
+    fun onLoginSuccess(message:String)
+    fun onLoginError(message:String)
+
+    //fun onLoginResult(message:String)
+
+}
