+++ app/src/main/java/com/genericsl/view/ui/fragment/gallery/GalleryFragment.kt
@@ -1,0 +1,31 @@
+package com.genericsl.view.ui.fragment.gallery
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import com.genericsl.R
+
+class GalleryFragment : Fragment() {
+
+    private lateinit var galleryViewModel: GalleryViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        galleryViewModel =
+            ViewModelProviders.of(this).get(GalleryViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
+        val textView: TextView = root.findViewById(R.id.text_gallery)
+        galleryViewModel.text.observe(this, Observer {
+            textView.text = it
+        })
+        return root
+    }
+}
