+++ app/src/main/java/com/genericsl/view/ui/fragment/gallery/GalleryViewModel.kt
@@ -1,0 +1,13 @@
+package com.genericsl.view.ui.fragment.gallery
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class GalleryViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is gallery Fragment"
+    }
+    val text: LiveData<String> = _text
+}
