+++ app/src/main/java/com/genericsl/view/ui/fragment/send/SendFragment.kt
@@ -1,0 +1,31 @@
+package com.genericsl.view.ui.fragment.send
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProviders
+import com.genericsl.R
+
+class SendFragment : Fragment() {
+
+    private lateinit var sendViewModel: SendViewModel
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        sendViewModel =
+            ViewModelProviders.of(this).get(SendViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_send, container, false)
+        val textView: TextView = root.findViewById(R.id.text_send)
+        sendViewModel.text.observe(this, Observer {
+            textView.text = it
+        })
+        return root
+    }
+}
