+++ app/src/main/java/com/genericsl/view/ui/fragment/share/ShareViewModel.kt
@@ -1,0 +1,13 @@
+package com.genericsl.view.ui.fragment.share
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class ShareViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is share Fragment"
+    }
+    val text: LiveData<String> = _text
+}
